+++ src/main/java/com/getpcpanel/hid/DeviceCommunicationHandlerFactory.java
@@ -1,6 +1,7 @@
 package com.getpcpanel.hid;
 
 import org.hid4java.HidDevice;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
@@ -10,10 +11,10 @@ import lombok.extern.log4j.Log4j2;
 @Service
 @RequiredArgsConstructor
 public class DeviceCommunicationHandlerFactory {
+    private final ApplicationEventPublisher eventPublisher;
     private final DeviceScanner deviceScanner;
-    private final InputInterpreter inputInterpreter;
 
     public DeviceCommunicationHandler build(String key, HidDevice device) {
-        return new DeviceCommunicationHandler(deviceScanner, inputInterpreter, key, device);
+        return new DeviceCommunicationHandler(deviceScanner, eventPublisher, key, device);
     }
 }